requirements.txt
langgraph~=0.2.50
langchain~=0.3.10
langchain-groq~=0.2.1
langchain-community~=0.3.10
langchain-huggingface~=0.1.0
faiss-cpu~=1.9.0
sentence-transformers~=3.3.0
pydantic~=2.9
python-dotenv~=1.0
streamlit~=1.40
pytest~=8.3
